requirements.txt
Flask==2.3.3
cryptography==41.0.4
requests==2.31.0
gunicorn==21.2.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
PyJWT==2.8.0
pyotp==2.9.0
qrcode[pil]==8.2
pytest==7.4.3
pytest-flask==1.3.0
